#324 refactor : 모바일 PortfolioHeader 아이콘사이즈 변경, Portfolio 테이블 비어있는 경우 수정
