+++ src/features/portfolio/components/PortfolioHolding/EmptyPortfolioHoldingTable.tsx
@@ -5,7 +5,7 @@ import { useBoolean } from "@fineants/demolition";
 import useResponsiveLayout from "@hooks/useResponsiveLayout";
 import designSystem from "@styles/designSystem";
 import styled from "styled-components";
-import PortfolioHoldingAddDialog from "./PortfolioHoldingAddDialogD";
+import PortfolioHoldingAddDialog from "./PortfolioHoldingAddDialog";
 
 export default function EmptyPortfolioHoldingTable() {
   const { isMobile } = useResponsiveLayout();
@@ -25,7 +25,10 @@ export default function EmptyPortfolioHoldingTable() {
           <div>종목을 추가하세요</div>
           <span>보유한 종목을 추가하여 포트폴리오 관리를 시작하세요</span>
         </TextBox>
-        <Button variant="primary" size="h32" onClick={onDialogOpen}>
+        <Button
+          variant="primary"
+          size={isMobile ? "h40" : "h32"}
+          onClick={onDialogOpen}>
           <Icon icon="add" size={16} color="white" />
           종목 추가
         </Button>
@@ -62,6 +65,12 @@ const TextBox = styled.div`
   flex-direction: column;
   gap: 8px;
   text-align: center;
+
+  > div {
+    font: ${designSystem.font.title3.font};
+    letter-spacing: ${designSystem.font.title3.letterSpacing};
+    color: ${designSystem.color.neutral.gray600};
+  }
 
   > span {
     font: ${designSystem.font.body3.font};
